1: Implemented interface.

+++ InventoryWebApi/Models/ItemQueryModel.cs
@@ -6,6 +6,7 @@ namespace InventoryWebApi.Models
         public int? Barcode { get; set; }
         public string? Name { get; set; }
         public string? Category { get; set; }
-        public int? Discount { get; set; }
+        public double? Discount { get; set; }
     }
 }
+
